feat: complete all PRD features - payment, notifications, promotions, reviews, customer, unified web app
Backend:
- Payment module: fixed enum/amount mismatches (BigDecimal->Integer), wired MidtransGatewayAdapter, webhook verification
- Notification module: removed duplicate DeliveryRecord, wired EmailPort+PushNotificationPort, fixed Kafka consumer, added producer
- Promotion module: built from scratch (Coupon, Campaign, Loyalty) with 5 entities, 5 repos, 3 services, 4 controllers
- Review module: fixed entity schema mismatch (rating/title/body), added ReviewResponse, provider review endpoints
- Customer module: CustomerService + CustomerController with profile CRUD, booking integration
- Blocked dates: BlockedDate entity + repository, provider CRUD, public endpoint for date picker
- Booking PIN: V20 migration (confirmation_pin + pin_verified), verifyPin endpoint, auto-generated 6-digit PIN

Frontend:
- Unified web_public app: merged customer + provider into single app with role-based routing
- Auth context: useAuth() with JWT decode, role-based navigation
- Profile completion: /profile/complete route for new customers
- Provider pages: Dashboard, Calendar, Services, Staff, Customers, Reports, Settings

Flutter:
- Fixed 37 deprecation warnings (withOpacity->withValues, textScaleFactor)
- Firebase wrapped in try-catch, services bootstrap safe

+++ services/platform_backend/src/main/java_root/id/dekat/payment/infrastructure/gateway/CreateTransactionRequest.java
@@ -3,7 +3,6 @@ package id.dekat.payment.infrastructure.gateway;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.List;
 
@@ -11,7 +10,7 @@ import java.util.List;
 @Builder
 public class CreateTransactionRequest {
     private String orderId;
-    private BigDecimal amount;
+    private Integer amount;
     private String currency;
     private String customerName;
     private String customerEmail;
@@ -25,7 +24,7 @@ public class CreateTransactionRequest {
     public static class OrderItem {
         private String id;
         private String name;
-        private BigDecimal price;
+        private Integer price;
         private Integer quantity;
     }
 }
